feat: Add PDF thumbnail support to static handler
This commit adds support for generating PDF thumbnails when requesting
static assets through the server. It checks the URL for a
'thumbnail=true' parameter and applies Cloudinary transformations to
extract the first page of the PDF as a JPEG. This is accomplished by
adding Cloudinary transformation parameters to the URL before fetching
the asset.